Writeup the module docs and move the protocol code
